began optimizing

+++ neptune/quad.vert
@@ -4,7 +4,7 @@ layout (location = 1) in vec2 in_texel;
 
 out vec2 texel;
 
-uniform vec2 size;
+uniform uvec2 size;
 
 void main() {
 	gl_Position = vec4(in_pos, 0, 1);
